# pkg/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zeroth API
  version: 0.0.0
  description: |
    Local control-plane API for Zeroth. Stage 1 is single-player and binds
    locally at 127.0.0.1:8420 by default (ZEROTH_ADDR or zerothd --addr).
    This spec is the canonical contract; Go stubs and the TypeScript client
    are generated from it, not hand-written.

    A flow that cannot be expressed through this API ships on neither the
    web UI nor the CLI. That is what keeps a later Slack surface from becoming
    a second implementation of everything.

    Plan lifecycle is draft, automatic cross-exam, approve (or request-changes,
    or branch), then apply. Apply is a client operation. Cross-exam is not:
    the daemon runs it after every new draft, including drafts produced by
    request-changes. There is no operator re-run trigger.
  license:
    name: MIT
    identifier: MIT
  contact:
    url: https://github.com/avivl/zeroth
servers:
  - url: http://127.0.0.1:8420
    description: |
      Default zerothd bind. Override with ZEROTH_ADDR or zerothd --addr.
      Stage 1 is loopback only; this is not a remote control plane.
...
paths:
...
  /runs:
    get:
      tags: [runs]
      summary: List runs
      operationId: listRuns
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          description: If set, only runs in this status.
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: agent_id
          in: query
          description: If set, only runs for this agent.
          schema:
            $ref: "#/components/schemas/AgentID"
      responses:
        "200":
          description: A page of runs, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
    post:
      tags: [runs]
      summary: Start a run
      description: |
        Stage 1 has exactly one sandbox driver (Docker). It is not
        user-selectable and is not a field on this request.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  "/runs/{id}/steer":
    post:
      tags: [runs]
      summary: Steer a run
      description: Inject operator guidance into a live run. Does not skip plan-then-apply.
      operationId: steerRun
      parameters:
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SteerRequest"
      responses:
        "200":
          description: Steer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  "/runs/{id}/background":
    post:
      tags: [runs]
      summary: Background a run
      description: Keep the run working without a foreground session. The operator can still steer, review plans, and read events.
      operationId: backgroundRun
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run is backgrounded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  "/runs/{id}/foreground":
    post:
      tags: [runs]
      summary: Foreground a run
      description: Bring a backgrounded run back to a foreground session. Mirrors POST /runs/{id}/background.
      operationId: foregroundRun
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run is in the foreground
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /agents:
    get:
      tags: [agents]
      summary: List agents
      operationId: listAgents
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
  "/agents/{id}":
    get:
      tags: [agents]
      summary: Get an agent
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/AgentID"
      responses:
        "200":
          description: The agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [agents]
      summary: Patch agent config
      description: |
        Config changes are themselves signed audited actions. The daemon
        records and signs the patch; the client does not send a signature.

        A tier change is an explicit operator action. Stage 1 has no
        auto-promotion. Leases are not set here; they are minted during
        apply. Read current leases at GET /agents/{id}/leases.
      operationId: patchAgent
      parameters:
        - $ref: "#/components/parameters/AgentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPatch"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /audit:
    get:
      tags: [audit]
      summary: List audit records
      operationId: listAudit
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: resource_type
          in: query
          schema:
            $ref: "#/components/schemas/AuditResourceType"
        - name: resource_id
          in: query
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: A page of audit records, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditList"
  "/audit/{id}/verify":
    post:
      tags: [audit]
      summary: Verify an audit record
      description: |
        Re-check the record's secp256k1 Schnorr signature (ADR-Z-0007) against
        the append-only key registry. Does not rewrite the log. Offline chain
        verification is `zeroth verify <run-id>` against the SQLite file.
      operationId: verifyAudit
      parameters:
        - $ref: "#/components/parameters/AuditID"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditVerification"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque page cursor from a previous next_cursor.
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      description: Page size. The daemon defaults this when omitted.
      schema:
        type: integer
        minimum: 1
        maximum: 200
    RunID:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/RunID"
...
    AgentID:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/AgentID"
...
    AuditID:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/AuditID"
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Policy denied this action (missing grant, expired lease, or unknown scope).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No resource with that id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The resource is not in a state that allows this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          minLength: 1
          description: Machine-readable error code.
        message:
          type: string
          minLength: 1
          description: Human-readable explanation.
    RunID:
      type: string
      minLength: 1
      description: Opaque run id. Not interchangeable with other id kinds.
    PlanID:
      type: string
      minLength: 1
      description: Opaque plan id. Not interchangeable with other id kinds.
    AgentID:
      type: string
      minLength: 1
      description: Opaque agent id. Not interchangeable with other id kinds.
...
    AuditID:
      type: string
      minLength: 1
      description: Opaque audit record id. Not interchangeable with other id kinds.
...
    LeaseID:
      type: string
      minLength: 1
      description: Opaque lease id. Not interchangeable with other id kinds.
...
    RunStatus:
      type: string
      description: Lifecycle of a run. Unknown values must be treated as non-terminal.
      enum:
        - pending
        - running
        - waiting_approval
        - backgrounded
        - stopped
        - completed
        - failed
        - cancelled
...
    AgentStatus:
      type: string
      enum:
        - ready
        - busy
        - disabled
...
    AuditResourceType:
      type: string
      enum:
        - run
        - plan
        - agent
        - memory
        - checkpoint
        - lease
...
    Run:
      type: object
      additionalProperties: false
      required: [id, agent_id, status, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/RunID"
        agent_id:
          $ref: "#/components/schemas/AgentID"
        plan_id:
          $ref: "#/components/schemas/PlanID"
        status:
          $ref: "#/components/schemas/RunStatus"
        prompt:
          type: string
          description: Operator prompt that started the run, when one was given.
        tracker_ref:
          type: string
          minLength: 1
          description: Optional tracker issue id (for example a Linear issue id). This API does not proxy the tracker.
        workspace_source:
          $ref: "#/components/schemas/WorkspaceSource"
        autonomy_tier:
          type: string
          minLength: 1
          description: Earned autonomy tier for this run. The kernel interprets the value; unknown tiers deny by default.
        created_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp in UTC.
        updated_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp in UTC.
        finished_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp in UTC.
    RunList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        next_cursor:
          type: string
          minLength: 1
          description: Omitted when there is no next page.
    WorkspaceSource:
      type: object
      additionalProperties: false
      required: [repo]
      properties:
        repo:
          type: string
          minLength: 1
          description: Repository URL or local path to check out for the run.
        ref:
          type: string
          minLength: 1
          description: Git ref (branch, tag, or SHA). The daemon default applies when omitted.
    CreateRunRequest:
      type: object
      additionalProperties: false
      required: [agent_id]
      properties:
        agent_id:
          $ref: "#/components/schemas/AgentID"
        tracker_ref:
          type: string
          minLength: 1
          description: Optional tracker issue id to attach to the run.
        prompt:
          type: string
          description: Optional operator prompt. Not required to create the run.
        workspace_source:
          $ref: "#/components/schemas/WorkspaceSource"
    SteerRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          description: Operator guidance to inject into the live run.
...
    Agent:
      type: object
      additionalProperties: false
      required: [id, name, harness, status, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/AgentID"
        name:
          type: string
          minLength: 1
        harness:
          type: string
          minLength: 1
          description: Harness driver name. Stage 1 is claudecode.
        status:
          $ref: "#/components/schemas/AgentStatus"
        model:
          type: string
          minLength: 1
        tools:
          type: array
          items:
            type: string
            minLength: 1
          description: Tool names this agent may be offered. Policy still denies by default; this is not a grant.
        autonomy_tier:
          type: string
          minLength: 1
          description: Default autonomy tier for new runs of this agent. Unknown tiers deny by default.
        created_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp in UTC.
        updated_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp in UTC.
    AgentPatch:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        tools:
          type: array
          items:
            type: string
            minLength: 1
        autonomy_tier:
          type: string
          minLength: 1
          description: Explicit operator tier change. Stage 1 does not auto-promote.
    AgentList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        next_cursor:
          type: string
          minLength: 1
...
    AuditRecord:
      type: object
      additionalProperties: false
      required: [id, action, resource_type, resource_id, signature, created_at]
      properties:
        id:
          $ref: "#/components/schemas/AuditID"
        action:
          type: string
          minLength: 1
          description: What happened (open string; clients must tolerate unknown actions).
        resource_type:
          $ref: "#/components/schemas/AuditResourceType"
        resource_id:
          type: string
          minLength: 1
        actor:
          type: string
          minLength: 1
          description: Local operator identity used when signing.
        target:
          type: string
          minLength: 1
          description: Resource the action applied to.
        plan_hash:
          type: string
          minLength: 1
          description: Canonical plan hash when the action is plan-scoped.
        precondition:
          type: string
        postcondition:
          type: string
        lease_id:
          $ref: "#/components/schemas/LeaseID"
        approver:
          type: string
          minLength: 1
        agent_id:
          $ref: "#/components/schemas/AgentID"
        run_id:
          $ref: "#/components/schemas/RunID"
        agent_pubkey:
          type: string
          minLength: 64
          maxLength: 64
          description: Hex-encoded BIP-340 x-only public key that produced signature.
        prev_hash:
          type: string
          description: Hex SHA-256 of the previous record (payload plus signature). Empty for genesis.
        hash:
          type: string
          minLength: 1
          description: Hex SHA-256 of this record's payload plus signature. The next row's prev_hash.
        signature:
          type: string
          minLength: 1
          description: secp256k1 Schnorr signature, hex-encoded (ADR-Z-0007). Not a credential.
        created_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp in UTC.
    AuditList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditRecord"
        next_cursor:
          type: string
          minLength: 1
    AuditVerification:
      type: object
      additionalProperties: false
      required: [id, valid]
      properties:
        id:
          $ref: "#/components/schemas/AuditID"
        valid:
          type: boolean
        reason:
          type: string
          description: Set when valid is false.
...